add: validation, update views, update fillable

+++ app/Http/Controllers/Admin/ApartmentController.php
@@ -33,17 +33,40 @@ class ApartmentController extends Controller
      */
     public function store(StoreApartmentRequest $request)
     {
+
+        $request->validate([
+            'description' => 'required|max:500',
+            'rooms' => 'required|numeric',
+            'beds' => 'required|numeric',
+            'bathrooms' => 'required|numeric',
+            'square_meters' => 'required|numeric',
+            'street_name' => 'required|max:255',
+            'street_number' => 'required|max:255',
+            'city' => 'required|max:255',
+            'postal_code' => 'required|max:255',
+            'cover_image' => 'string'
+        ]);
+
         $data = $request->all();
 
-        // Geocode apartment address to latitude and longitude using tomtom api
-        $response = Http::withOptions(['verify' => false])
-            ->get('https://api.tomtom.com/search/2/search/' . $data['address'] . '.json?key=qD5AjlcGdPMFjUKdDAYqT7xYi3yIRo3c');
+        $full_address =
+            $data['street_name'] . ', ' .
+            $data['street_number'] . ', ' .
+            $data['city'] . ', ' .
+            $data['postal_code'];
+
+        $base_url = "https://api.tomtom.com/search/2/search/";
+        $api_key = "?key=qD5AjlcGdPMFjUKdDAYqT7xYi3yIRo3c";
+        $responseFormat = ".json";
+        $query_url = $base_url . $full_address . $responseFormat . $api_key;
+
+        $response = Http::withOptions(['verify' => false])->get($query_url);
         $results = $response->json()["results"];
         $data["latitude"] = $results[0]["position"]["lat"];
         $data["longitude"] = $results[0]["position"]["lon"];
 
         if ($request->hasFile('cover_image')) {
-            $path = Storage::put('cover_images',$request->cover_image);
+            $path = Storage::put('cover_images', $request->cover_image);
             $data['cover_image'] = $path;
         }
 
@@ -80,8 +103,11 @@ class ApartmentController extends Controller
             'beds' => 'required|numeric',
             'bathrooms' => 'required|numeric',
             'square_meters' => 'required|numeric',
-            'address' => 'required|max:255',
-            'cover_image' => 'required|string'
+            'street_name' => 'required|max:255',
+            'street_number' => 'required|max:255',
+            'city' => 'required|max:255',
+            'postal_code' => 'required|max:255',
+            'cover_image' => 'string'
         ]);
 
         $data = $request->all();
@@ -96,7 +122,9 @@ class ApartmentController extends Controller
      */
     public function destroy(Apartment $apartment)
     {
-        Storage::delete($apartment->cover_image);
+        if ($apartment->cover_image) {
+            Storage::delete($apartment->cover_image);
+        }
 
         $apartment->delete();
         return redirect()->route('admin.apartments.index');
